move contents of buildUI to special GUI class
